+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Produce `arm64` binaries with Nuitka by using `macos-14` runners.
+
 ## [3.0.0 (2024-02-17)](https://github.com/kdeldycke/workflows/compare/v2.26.6...v3.0.0)
 
 - Start replacing `pip` invocations by `uv`.